requirements.txt
weaviate-client>=3.0.0
requests>=2.25.0
PyGithub>=1.55
python-dotenv>=0.19.0

